Refactored dialog component into a root and panel
