+++ v1-migration-folder/packages/react-components/src/dialog/dialog.stories.tsx
@@ -1,26 +1,21 @@
 import "./dialog.stories.css";
 
 import type { Meta, StoryObj } from "@storybook/react";
-import { Dialog, type DialogRootProps } from "./dialog";
+import { DialogRoot, type DialogRootProps } from "./root";
+import { DialogPanel } from "./dialog";
 
 const meta: Meta<DialogRootProps> = {
   title: "components/Dialog",
-  component: Dialog,
-  args: {
-    open: false,
-    className: "dialog",
-    timeEnter: 100,
-    timeExit: 100,
-  },
+  component: DialogRoot,
 };
 
 export default meta;
 
 function WithChildren(props: DialogRootProps) {
   return (
-    <Dialog {...props}>
-      <div>Run Me</div>
-    </Dialog>
+    <DialogRoot {...props}>
+      <DialogPanel className="dialog">This is my dialog panel</DialogPanel>
+    </DialogRoot>
   );
 }
