+++ v1-migration-folder/packages/react-components/src/dialog/dialog.tsx
@@ -1,61 +1,12 @@
-import { createPortal } from "react-dom";
-import { useMemo, useState, type CSSProperties, type PropsWithChildren } from "react";
-import { useManagedDialog } from "./use-managed-dialog";
-import type { OpenState } from "./use-transition-effect";
+import { forwardRef, type JSX } from "react";
+import { useDialog } from "./root.use-dialog";
+import { useForkRef } from "@1771technologies/lytenyte-react-hooks";
 
-/**
- * <root>
- *  <trigger>
- *  <portal>
- *    <panel>
- *      <title />
- *      <description />
- *    </panel>
- *  </portal>
- * </root>
- */
+export const DialogPanel = forwardRef<HTMLDialogElement, JSX.IntrinsicElements["dialog"]>(
+  function DialogPanel(props, forwarded) {
+    const { state, dialogRef } = useDialog();
+    const mergedRefs = useForkRef(forwarded, dialogRef);
 
-export interface DialogRootProps {
-  readonly open?: boolean;
-  readonly onOpenChange?: (b: boolean) => void;
-  readonly timeEnter?: number;
-  readonly timeExit?: number;
-
-  readonly className?: string;
-  readonly style?: CSSProperties | ((state: OpenState) => CSSProperties);
-}
-
-export function Dialog({
-  open,
-  onOpenChange,
-  children,
-  style,
-  timeEnter = 0,
-  timeExit = 0,
-  ...props
-}: PropsWithChildren<DialogRootProps>) {
-  const [localOpen, setLocalOpen] = useState(false);
-
-  const { shouldMount, ref, state } = useManagedDialog(
-    open ?? localOpen,
-    onOpenChange ?? setLocalOpen,
-    timeEnter,
-    timeExit,
-  );
-
-  const finalStyle = useMemo(() => {
-    if (!style) return;
-
-    if (typeof style === "function") return style(state);
-    return style;
-  }, [state, style]);
-
-  return createPortal(
-    shouldMount && (
-      <dialog {...props} data-dialog-state={state} ref={ref} style={finalStyle}>
-        {children}
-      </dialog>
-    ),
-    document.body,
-  );
-}
+    return <dialog {...props} data-dialog-state={state} ref={mergedRefs}></dialog>;
+  },
+);
