+++ v1-migration-folder/packages/react-components/src/dialog/index.ts
@@ -1,0 +1,21 @@
+import { DialogPanel } from "./dialog.js";
+import { DialogRoot } from "./root.js";
+
+export const Dialog = {
+  Root: DialogRoot,
+  Panel: DialogPanel,
+};
+
+export type { DialogApi, DialogRootProps } from "./root.js";
+
+/**
+ * <root>  done
+ *  <trigger>
+ *  <portal>
+ *    <panel> done
+ *      <title />
+ *      <description />
+ *    </panel> done
+ *  </portal>
+ * </root> done
+ */
